backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
anthropic==0.7.8
requests==2.31.0
python-dotenv==1.0.0
networkx==3.2.1
pandas==2.1.4
aiohttp==3.9.1
numpy==1.25.2
yfinance==0.2.28
newsapi-python==0.2.7
python-multipart==0.0.6
beautifulsoup4==4.12.2
lxml==4.9.3
python-dateutil==2.8.2
networkx==3.2.1
